Switch to having Flower use a config file, enable OAuth
Google OAuth requires an additional 4 command line args or environment
variables, so it was time for Flower to get a proper config file.

Folds in the address and port settings from the command line.

The broker configuration remains a command line argument because it is
consumed by Celery's configuration parser, not Flower's configuration
parser.